Common header updated and ThemeInputsConfig
